feat: Implement OpenAPI utilities and schema generation
- Added openapi-utils.ts for handling OpenAPI schema object checks and response reference resolution.
- Introduced operation-utils.ts to iterate through operations in an OpenAPI document.
- Created package-generator.ts to generate package.json for the output.
- Developed request-body-resolver.ts to resolve request body references in OpenAPI documents.
- Implemented schema-conflict-resolver.ts to rename conflicting schemas and prevent identifier collisions.
- Added schema-extractor.ts to extract request and response schemas from operations.
- Created schema-generation-coordinator.ts to manage the generation of component, request, and response schemas.
