+++ apps/craft/src/core-generator/index.ts
@@ -1,33 +1,21 @@
 /* eslint-disable no-console */
-import type {
-  OpenAPIObject,
-  OperationObject,
-  RequestBodyObject,
-  ResponseObject,
-  SchemaObject,
-} from "openapi3-ts/oas31";
+import type { OpenAPIObject } from "openapi3-ts/oas31";
 
 import $RefParser from "@apidevtools/json-schema-ref-parser";
-import assert from "assert";
 import { promises as fs } from "fs";
-import { isReferenceObject } from "openapi3-ts/oas31";
-import pLimit from "p-limit";
-import path from "path";
 
 import { generateOperations } from "../client-generator/index.js";
 import { applyGeneratedOperationIds } from "../operation-id-generator/index.js";
-import {
-  analyzeSchemaForRecursion,
-  createRecursiveContext,
-  generateRecursiveSchemaFile,
-  generateRequestSchemaFile,
-  generateResponseSchemaFile,
-  generateSchemaFile,
-} from "../schema-generator/index.js";
-import { sanitizeIdentifier } from "../schema-generator/utils.js";
 import { generateServerOperations } from "../server-generator/index.js";
 import { convertToOpenAPI31 } from "./converter.js";
+import { createPackageJson } from "./package-generator.js";
 import { parseOpenAPI } from "./parser.js";
+import { resolveRequestBodies } from "./request-body-resolver.js";
+import {
+  renameConflictingSchemas,
+  renameSanitizationConflictingSchemas,
+} from "./schema-conflict-resolver.js";
+import { generateAllSchemas } from "./schema-generation-coordinator.js";
 
 const DEFAULT_CONCURRENCY = 10;
 
@@ -63,33 +51,6 @@ export interface GenerationOptions {
    */
   strictValidation?: boolean;
 }
-
-/**
- * Options for component schema promise creation
- */
-interface ComponentSchemaOptions {
-  context: SchemaGenerationContext;
-  description?: string;
-  originalSchemaName?: string;
-  recursiveContext: ReturnType<typeof createRecursiveContext>;
-  schema: SchemaObject;
-  schemaName: string;
-  strictValidation: boolean;
-}
-
-interface SchemaGenerationContext {
-  generateServer: boolean;
-  limit: ReturnType<typeof pLimit>;
-  openApiDoc: OpenAPIObject;
-  schemasDir: string;
-  strictValidation: boolean;
-}
-
-type SchemaGeneratorFunction<T = SchemaObject> = (
-  name: string,
-  schema: T,
-  options: { strictValidation: boolean },
-) => Promise<{ content: string; fileName: string }>;
 
 /**
  * Generates TypeScript schemas and optional API client from OpenAPI specification
@@ -128,349 +89,6 @@ export async function generate(options: GenerationOptions): Promise<void> {
 }
 
 /**
- * Resolves a response reference within an OpenAPI document
- */
-export function resolveResponseReference(
-  ref: string,
-  doc: OpenAPIObject,
-): ResponseObject | undefined {
-  if (!ref.startsWith("#/components/responses/")) {
-    return undefined;
-  }
-
-  const responseName = ref.replace("#/components/responses/", "");
-  const response = doc.components?.responses?.[responseName];
-
-  if (!response) {
-    return undefined;
-  }
-
-  // The resolved response should be a ResponseObject, not a ReferenceObject
-  // If it's still a reference, we'd need recursive resolution, but OpenAPI bundling
-  // should have resolved this already
-  if (isReferenceObject(response)) {
-    console.warn(
-      `⚠️ Nested response reference not resolved: ${ref} -> ${response.$ref}`,
-    );
-    return undefined;
-  }
-
-  return response;
-}
-
-/**
- * Creates a schema generation promise for a single schema variant
- */
-function createComponentSchemaPromise(
-  options: ComponentSchemaOptions,
-): Promise<void> {
-  const {
-    context,
-    description,
-    originalSchemaName,
-    recursiveContext,
-    schema,
-    schemaName,
-    strictValidation,
-  } = options;
-  const isRecursive = recursiveContext.recursiveSchemas.has(schemaName);
-
-  return context.limit(async () => {
-    const generationPromise = isRecursive
-      ? generateRecursiveSchemaFile({
-          description,
-          name: schemaName,
-          originalSchemaName: originalSchemaName || schemaName,
-          recursiveContext,
-          schema,
-          strictValidation,
-        })
-      : generateSchemaFile(schemaName, schema, description, {
-          originalSchemaName,
-          recursiveContext,
-          strictValidation,
-        });
-
-    const schemaFile = await generationPromise;
-    const filePath = path.join(context.schemasDir, schemaFile.fileName);
-    return await fs.writeFile(filePath, schemaFile.content);
-  });
-}
-
-/**
- * Creates the package.json file for the generated output
- */
-async function createPackageJson(output: string): Promise<void> {
-  const packageJsonContent = {
-    dependencies: {
-      zod: "^4.0.0",
-    },
-    devDependencies: {
-      "@types/node": "^24.3.1",
-      typescript: "^5.4.5",
-    },
-    name: "generated-client",
-    scripts: {
-      build:
-        "tsc --outDir ./dist --rootDir . --moduleResolution NodeNext --module NodeNext --target ES2022 --lib es2022 --strict --esModuleInterop --skipLibCheck --allowSyntheticDefaultImports --resolveJsonModule --forceConsistentCasingInFileNames --noEmitOnError false schemas/*.ts client/*.ts server/*.ts",
-    },
-    type: "module",
-    version: "0.1.0",
-  };
-  const packageJsonPath = path.join(output, "package.json");
-  await fs.writeFile(
-    packageJsonPath,
-    JSON.stringify(packageJsonContent, null, 2),
-  );
-}
-
-/**
- * Creates schema generation promises for both regular and strict variants
- */
-function createSchemaGenerationPromises<T = SchemaObject>(
-  schemaMap: Map<string, T>,
-  context: SchemaGenerationContext,
-  generatorFn: SchemaGeneratorFunction<T>,
-): Promise<void>[] {
-  const promises: Promise<void>[] = [];
-
-  for (const [name, schema] of schemaMap) {
-    // Generate regular schema
-    const promise = context.limit(() =>
-      generatorFn(name, schema, {
-        strictValidation: context.strictValidation,
-      }).then((schemaFile) => {
-        const filePath = path.join(context.schemasDir, schemaFile.fileName);
-        return fs.writeFile(filePath, schemaFile.content);
-      }),
-    );
-    promises.push(promise);
-
-    // Generate strict schema for server when generateServer is enabled
-    if (context.generateServer) {
-      const strictPromise = context.limit(() =>
-        generatorFn(`${name}Strict`, schema, {
-          strictValidation: true,
-        }).then((schemaFile) => {
-          const filePath = path.join(context.schemasDir, schemaFile.fileName);
-          return fs.writeFile(filePath, schemaFile.content);
-        }),
-      );
-      promises.push(strictPromise);
-    }
-  }
-
-  return promises;
-}
-
-/**
- * Extracts request schemas from operations for inline request body schemas
- *
- * @example
- * ```javascript
- * const openApiDoc = {
- *   paths: {
- *     '/users': {
- *       post: {
- *         operationId: 'createUser',
- *         requestBody: {
- *           content: {
- *             'application/json': {
- *               schema: {
- *                 type: 'object',
- *                 properties: { name: { type: 'string' } }
- *               }
- *             }
- *           }
- *         }
- *       }
- *     }
- *   }
- * };
- *
- * const schemas = extractRequestSchemas(openApiDoc);
- * // Result: Map with entry 'CreateUserRequest' -> schema object
- * ```
- */
-function extractRequestSchemas(
-  openApiDoc: OpenAPIObject,
-): Map<string, SchemaObject> {
-  const requestSchemas = new Map<string, SchemaObject>();
-
-  forEachOperation(openApiDoc, (operation) => {
-    if (!operation.requestBody) return;
-
-    // Handle both direct RequestBodyObject and ReferenceObject
-    let requestBody: RequestBodyObject;
-    if (isReferenceObject(operation.requestBody)) {
-      // Skip reference objects for now - we only want inline schemas
-      return;
-    } else {
-      requestBody = operation.requestBody;
-    }
-
-    const supportedContentTypes = [
-      "application/json",
-      "multipart/form-data",
-      "application/x-www-form-urlencoded",
-      "application/octet-stream",
-    ];
-
-    assert(operation.operationId, "Operation ID is missing");
-
-    for (const contentType of supportedContentTypes) {
-      const content = requestBody.content?.[contentType];
-      if (content?.schema && !isReferenceObject(content.schema)) {
-        // Only extract inline schemas, not $ref schemas
-        const requestTypeName = `${sanitizeIdentifier(operation.operationId)}Request`;
-        requestSchemas.set(requestTypeName, content.schema);
-        break; // Only process the first matching content type
-      }
-    }
-  });
-
-  return requestSchemas;
-}
-
-/**
- * Extracts response schemas from operations for inline response schemas
- *
- * @example
- * ```javascript
- * const openApiDoc = {
- *   paths: {
- *     '/users/{id}': {
- *       get: {
- *         operationId: 'getUser',
- *         responses: {
- *           '200': {
- *             content: {
- *               'application/json': {
- *                 schema: {
- *                   type: 'object',
- *                   properties: { id: { type: 'string' }, name: { type: 'string' } }
- *                 }
- *               }
- *             }
- *           }
- *         }
- *       }
- *     }
- *   }
- * };
- *
- * const schemas = extractResponseSchemas(openApiDoc);
- * // Result: Map with entry 'GetUser200Response' -> schema object
- * ```
- */
-function extractResponseSchemas(
-  openApiDoc: OpenAPIObject,
-): Map<string, SchemaObject> {
-  const responseSchemas = new Map<string, SchemaObject>();
-
-  forEachOperation(openApiDoc, (operation) => {
-    if (!operation.responses) return;
-
-    const operationId = operation.operationId;
-
-    assert(operationId, "Operation ID is missing");
-
-    for (const [statusCode, response] of Object.entries(operation.responses)) {
-      // Process all status codes, including default responses
-
-      // Handle both direct ResponseObject and ReferenceObject
-      let responseObj: ResponseObject;
-      if (isReferenceObject(response)) {
-        // Resolve the response reference
-        const resolved = resolveResponseReference(response.$ref, openApiDoc);
-        if (!resolved) {
-          console.warn(
-            `⚠️ Could not resolve response reference: ${response.$ref}`,
-          );
-          continue;
-        }
-        responseObj = resolved;
-      } else {
-        responseObj = response;
-      }
-
-      if (!responseObj.content) continue;
-
-      // Check for various content types
-      const supportedContentTypes = [
-        "application/json",
-        "application/problem+json",
-        "application/octet-stream",
-        "multipart/form-data",
-        "application/pdf",
-      ];
-
-      for (const contentType of Object.keys(responseObj.content)) {
-        if (
-          supportedContentTypes.includes(contentType) ||
-          contentType.includes("+json")
-        ) {
-          const content = responseObj.content[contentType];
-          if (content?.schema && !isReferenceObject(content.schema)) {
-            // Only extract inline schemas, not $ref schemas
-            const sanitizedOperationId = sanitizeIdentifier(operationId);
-            const suffix =
-              statusCode === "default"
-                ? "DefaultResponse"
-                : `${statusCode}Response`;
-            const responseTypeName = `${sanitizedOperationId.charAt(0).toUpperCase() + sanitizedOperationId.slice(1)}${suffix}`;
-            responseSchemas.set(responseTypeName, content.schema);
-          }
-          break; // Only process the first matching content type
-        }
-      }
-    }
-  });
-
-  return responseSchemas;
-}
-
-/**
- * Common utility to iterate through all operations in an OpenAPI document.
- * Works with all operations, regardless of whether they have operationId
- */
-function forEachOperation(
-  openApiDoc: OpenAPIObject,
-  callback: (
-    operation: OperationObject,
-    method: string,
-    pathKey: string,
-  ) => void,
-): void {
-  if (!openApiDoc.paths) {
-    return;
-  }
-
-  for (const [pathKey, pathItem] of Object.entries(openApiDoc.paths)) {
-    const pathItemObj = pathItem;
-
-    // Define the HTTP methods we support with their corresponding operations
-    const httpMethods: {
-      method: string;
-      operation: OperationObject | undefined;
-    }[] = [
-      { method: "get", operation: pathItemObj.get },
-      { method: "post", operation: pathItemObj.post },
-      { method: "put", operation: pathItemObj.put },
-      { method: "delete", operation: pathItemObj.delete },
-      { method: "patch", operation: pathItemObj.patch },
-    ];
-
-    for (const { method, operation } of httpMethods) {
-      if (operation) {
-        // OperationId is generated if missing
-        callback(operation, method, pathKey);
-      }
-    }
-  }
-}
-
-/**
  * Generates all operations (client and/or server)
  */
 async function generateAllOperations(
@@ -488,156 +106,6 @@ async function generateAllOperations(
     await generateServerOperations(openApiDoc, output, concurrency);
     console.log("✅ Server operations generated successfully");
   }
-}
-
-/**
- * Generates all schemas (component, request, and response schemas)
- */
-async function generateAllSchemas(
-  openApiDoc: OpenAPIObject,
-  output: string,
-  concurrency: number,
-  strictValidation: boolean,
-  generateServer: boolean,
-): Promise<void> {
-  if (!openApiDoc.components?.schemas) {
-    return;
-  }
-
-  const schemasDir = path.join(output, "schemas");
-  await fs.mkdir(schemasDir, { recursive: true });
-
-  const limit = pLimit(concurrency);
-  const context: SchemaGenerationContext = {
-    generateServer,
-    limit,
-    openApiDoc,
-    schemasDir,
-    strictValidation,
-  };
-
-  const schemaGenerationPromises: Promise<void>[] = [
-    // Generate schemas from components/schemas
-    ...generateComponentSchemas(context),
-    // Generate request schemas from operations
-    ...createSchemaGenerationPromises(
-      extractRequestSchemas(openApiDoc),
-      context,
-      generateRequestSchemaFile,
-    ),
-    // Generate response schemas from operations
-    ...createSchemaGenerationPromises(
-      extractResponseSchemas(openApiDoc),
-      context,
-      generateResponseSchemaFile,
-    ),
-  ];
-
-  await Promise.all(schemaGenerationPromises);
-  console.log("✅ Schemas generated successfully");
-}
-
-/**
- * Generates component schemas from the OpenAPI document
- */
-function generateComponentSchemas(
-  context: SchemaGenerationContext,
-): Promise<void>[] {
-  const promises: Promise<void>[] = [];
-
-  if (!context.openApiDoc.components?.schemas) {
-    return promises;
-  }
-
-  // Create a shared recursive context for all schemas
-  const recursiveContext = createRecursiveContext();
-
-  // First pass: analyze all schemas for recursive patterns
-  for (const [name, schema] of Object.entries(
-    context.openApiDoc.components.schemas,
-  )) {
-    if (!isPlainSchemaObject(schema)) {
-      continue;
-    }
-
-    const sanitizedName = sanitizeIdentifier(name);
-
-    // Analyze for recursion and update context
-    const isRecursive = analyzeSchemaForRecursion(name, schema);
-    if (isRecursive) {
-      recursiveContext.recursiveSchemas.add(sanitizedName);
-      // Also add the strict variant if server generation is enabled
-      if (context.generateServer) {
-        recursiveContext.recursiveSchemas.add(`${sanitizedName}Strict`);
-      }
-    }
-  }
-
-  // Second pass: generate schema files with recursive context
-  for (const [name, schema] of Object.entries(
-    context.openApiDoc.components.schemas,
-  )) {
-    if (!isPlainSchemaObject(schema)) {
-      console.warn(
-        `⚠️ Skipping ${name}: not a plain OpenAPI schema object. Value:`,
-        schema,
-      );
-      continue;
-    }
-
-    const sanitizedName = sanitizeIdentifier(name);
-    const description = schema.description
-      ? schema.description.trim()
-      : undefined;
-
-    // Generate regular schema
-    const promise = createComponentSchemaPromise({
-      context,
-      description,
-      recursiveContext,
-      schema,
-      schemaName: sanitizedName,
-      strictValidation: context.strictValidation,
-    });
-    promises.push(promise);
-
-    // Generate strict schema for server when generateServer is enabled
-    if (context.generateServer) {
-      const strictPromise = createComponentSchemaPromise({
-        context,
-        description,
-        originalSchemaName: sanitizedName,
-        recursiveContext,
-        schema,
-        schemaName: `${sanitizedName}Strict`,
-        strictValidation: true,
-      });
-      promises.push(strictPromise);
-    }
-  }
-
-  return promises;
-}
-
-/**
- * Determines if an object is a plain OpenAPI schema object or reference object
- */
-function isPlainSchemaObject(obj: unknown): obj is SchemaObject {
-  if (!obj || typeof obj !== "object") return false;
-  // Check if it's a reference object (contains $ref)
-  if (isReferenceObject(obj)) {
-    return true;
-  }
-  // Must have at least one OpenAPI schema property
-  return (
-    "type" in obj ||
-    "allOf" in obj ||
-    "anyOf" in obj ||
-    "oneOf" in obj ||
-    "properties" in obj ||
-    "additionalProperties" in obj ||
-    "array" in obj
-  );
 }
 
 /**
@@ -707,286 +175,4 @@ async function parseAndPreprocessOpenAPI(
   }
 
   return openApiDoc;
-}
-
-/*
- * Renames schemas in components/schemas that conflict with:
- *  - Internal generator exported type names (e.g. ApiResponse)
- *  - Global / built-in JavaScript constructors or DOM-like types (Blob, Buffer, File, etc.)
- *  - Any explicitly reserved names defined below
- * The renaming strategy appends 'Schema' (and numeric suffix if needed) and updates every
- * $ref string pointing to the old schema name anywhere in the OpenAPI document.
- */
-function renameConflictingSchemas(openApiDoc: OpenAPIObject) {
-  if (!openApiDoc.components || !openApiDoc.components.schemas) return 0;
-  const schemas = openApiDoc.components.schemas;
-
-  const reservedNames = new Set<string>([
-    /* Reserved & built-ins (sorted) */
-    "ApiResponse",
-    "ApiResponseError",
-    "ApiResponseWithForcedParse",
-    "ApiResponseWithParse",
-    "Blob",
-    "Buffer",
-    "Date",
-    "Error",
-    "File",
-    "FormData",
-    "Headers",
-    "Map",
-    "Promise",
-    "ReadableStream",
-    "Request",
-    "Response",
-    "Set",
-    "TransformStream",
-    "URL",
-    "URLSearchParams",
-    "WeakMap",
-    "WeakSet",
-    "WritableStream",
-  ]);
-
-  const renameMap = new Map<string, string>();
-
-  for (const originalName of Object.keys(schemas)) {
-    if (!reservedNames.has(originalName)) continue;
-    let candidate = `${originalName}Schema`;
-    let counter = 2;
-    while (
-      schemas[candidate as keyof typeof schemas] ||
-      renameMap.has(candidate)
-    ) {
-      candidate = `${originalName}Schema${counter++}`;
-    }
-    renameMap.set(originalName, candidate);
-  }
-
-  if (renameMap.size === 0) return 0; /* Nothing to do */
-
-  /* Perform the actual renames (reconstruct object to avoid dynamic delete issues) */
-  if (renameMap.size) {
-    const rebuilt: Record<string, unknown> = {};
-    for (const [key, value] of Object.entries(schemas)) {
-      const newKey = renameMap.get(key) ?? key;
-      rebuilt[newKey] = value as unknown;
-    }
-    // Remove existing keys
-    for (const key of Object.keys(schemas)) {
-      // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-      delete (schemas as Record<string, unknown>)[key];
-    }
-    // Assign rebuilt
-    for (const [key, value] of Object.entries(rebuilt)) {
-      (schemas as Record<string, unknown>)[key] = value;
-    }
-  }
-
-  const refPrefix = "#/components/schemas/";
-
-  /* Walk entire document and rewrite $ref strings */
-  const visit = (node: unknown): void => {
-    if (!node) return;
-    if (Array.isArray(node)) {
-      for (const item of node) visit(item);
-      return;
-    }
-    if (typeof node === "object") {
-      const obj = node as Record<string, unknown>;
-      if (typeof obj.$ref === "string") {
-        const ref: string = obj.$ref;
-        if (ref.startsWith(refPrefix)) {
-          const name = ref.substring(refPrefix.length);
-          const renamed = renameMap.get(name);
-          if (renamed) obj.$ref = refPrefix + renamed;
-        }
-      }
-      for (const value of Object.values(obj)) visit(value);
-    }
-  };
-
-  visit(openApiDoc);
-  return renameMap.size;
-}
-
-/*
- * Renames schemas whose sanitized identifiers would collide with each other.
- * This prevents case-sensitivity conflicts when schemas like 'Catalog' and '_catalog'
- * both sanitize to similar identifiers, causing TypeScript import and filesystem issues.
- * The renaming strategy appends numeric suffixes and updates $ref pointers accordingly.
- */
-function renameSanitizationConflictingSchemas(
-  openApiDoc: OpenAPIObject,
-): number {
-  if (!openApiDoc.components || !openApiDoc.components.schemas) return 0;
-  const schemas = openApiDoc.components.schemas;
-
-  // Map from sanitized names (lowercase) to original schema names that would collide
-  const sanitizedToOriginals = new Map<string, string[]>();
-
-  // First pass: group original names by their sanitized equivalents (case-insensitive)
-  for (const originalName of Object.keys(schemas)) {
-    const sanitized = sanitizeIdentifier(originalName);
-    const sanitizedLower = sanitized.toLowerCase();
-
-    if (!sanitizedToOriginals.has(sanitizedLower)) {
-      sanitizedToOriginals.set(sanitizedLower, []);
-    }
-    const existingList = sanitizedToOriginals.get(sanitizedLower);
-    if (existingList) {
-      existingList.push(originalName);
-    }
-  }
-
-  // Find groups with more than one original name (collisions)
-  const collisionGroups = Array.from(sanitizedToOriginals.entries()).filter(
-    ([, originals]) => originals.length > 1,
-  );
-
-  if (collisionGroups.length === 0) return 0;
-
-  const renameMap = new Map<string, string>();
-
-  // For each collision group, rename all but the first schema
-  for (const [, originals] of collisionGroups) {
-    // Sort originals for deterministic behavior - keep the lexicographically first one unchanged
-    const sortedOriginals = [...originals].sort();
-
-    // Leave the first one unchanged, rename the rest
-    for (let i = 1; i < sortedOriginals.length; i++) {
-      const originalName = sortedOriginals[i];
-      const baseSanitized = sanitizeIdentifier(originalName);
-
-      // Find a unique name by appending numeric suffix
-      let candidate = `${baseSanitized}${i + 1}`;
-      let counter = i + 1;
-
-      // Ensure the candidate doesn't conflict with existing schemas or other renames
-      while (
-        schemas[candidate as keyof typeof schemas] ||
-        renameMap.has(candidate) ||
-        // Check if the new candidate would create another case-insensitive collision
-        Array.from(renameMap.values()).some(
-          (existing) => existing.toLowerCase() === candidate.toLowerCase(),
-        ) ||
-        Object.keys(schemas).some(
-          (existing) =>
-            sanitizeIdentifier(existing).toLowerCase() ===
-              candidate.toLowerCase() && !renameMap.has(existing),
-        )
-      ) {
-        counter++;
-        candidate = `${baseSanitized}${counter}`;
-      }
-
-      renameMap.set(originalName, candidate);
-    }
-  }
-
-  if (renameMap.size === 0) return 0;
-
-  // Perform the actual renames (reconstruct object to avoid dynamic delete issues)
-  const rebuilt: Record<string, unknown> = {};
-  for (const [key, value] of Object.entries(schemas)) {
-    const newKey = renameMap.get(key) ?? key;
-    rebuilt[newKey] = value as unknown;
-  }
-
-  // Remove existing keys
-  for (const key of Object.keys(schemas)) {
-    // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
-    delete (schemas as Record<string, unknown>)[key];
-  }
-
-  // Assign rebuilt
-  for (const [key, value] of Object.entries(rebuilt)) {
-    (schemas as Record<string, unknown>)[key] = value;
-  }
-
-  const refPrefix = "#/components/schemas/";
-
-  // Walk entire document and rewrite $ref strings
-  const visit = (node: unknown): void => {
-    if (!node) return;
-    if (Array.isArray(node)) {
-      for (const item of node) visit(item);
-      return;
-    }
-    if (typeof node === "object") {
-      const obj = node as Record<string, unknown>;
-      if (typeof obj.$ref === "string") {
-        const ref: string = obj.$ref;
-        if (ref.startsWith(refPrefix)) {
-          const name = ref.substring(refPrefix.length);
-          const renamed = renameMap.get(name);
-          if (renamed) obj.$ref = refPrefix + renamed;
-        }
-      }
-      for (const value of Object.values(obj)) visit(value);
-    }
-  };
-
-  visit(openApiDoc);
-  return renameMap.size;
-}
-
-/*
- * Resolves requestBodies references by replacing $ref pointers with inline content.
- * This preprocessing step allows the existing client and server generators to work
- * without modifications, as they only need to handle inline request body definitions.
- *
- * For each requestBodies reference found in operation requestBody, this function:
- * 1. Looks up the requestBody definition in components/requestBodies
- * 2. Replaces the $ref with the actual requestBody content
- * 3. Updates all references across the entire OpenAPI document
- */
-function resolveRequestBodies(openApiDoc: OpenAPIObject): number {
-  // Access requestBodies safely - extend components type for requestBodies support
-  const components = openApiDoc.components;
-  if (!components?.requestBodies) {
-    return 0; // No requestBodies to resolve
-  }
-
-  const requestBodies = components.requestBodies;
-  let resolvedCount = 0;
-
-  const refPrefix = "#/components/requestBodies/";
-
-  /* Walk entire document and replace requestBodies $ref with inline content */
-  const visit = (node: unknown): void => {
-    if (!node) return;
-    if (Array.isArray(node)) {
-      for (const item of node) visit(item);
-      return;
-    }
-    if (typeof node === "object") {
-      const obj = node as Record<string, unknown>;
-      if (isReferenceObject(obj)) {
-        const ref: string = obj.$ref;
-        if (ref.startsWith(refPrefix)) {
-          const requestBodyName = ref.substring(refPrefix.length);
-          const requestBody = requestBodies[requestBodyName];
-
-          if (requestBody) {
-            // Replace the $ref with the actual requestBody content
-            delete (obj as Record<string, unknown>).$ref;
-
-            // Copy all properties from the requestBody to the current object
-            for (const [key, value] of Object.entries(requestBody)) {
-              obj[key] = value;
-            }
-
-            resolvedCount++;
-          } else {
-            console.warn(`⚠️ Could not resolve requestBody reference: ${ref}`);
-          }
-        }
-      }
-      for (const value of Object.values(obj)) visit(value);
-    }
-  };
-
-  visit(openApiDoc);
-  return resolvedCount;
 }
